Remove Multiplicity from AttributeDefinition
The SAML2 specification does not make any mention
of multiplicity. Furthermore the functionality
was used sparsely if at all and should not have
been relied on as it was configuration based,
which could have changed causing a BC break in
and of itself as an Attribute could suddenly
cause exceptions to be thrown when configuration
changed or was overridden.