Enable POST-binding for signed AuthnRequests
The second factor only single sign-on endpoint needs to support POST
binding to support AuthnRequests originating from ADFS. To facilitate
this change, the PostBinding needs to be able to receive signed
AuthnRequest from POST requests.

To determine the difference between Redirect and Post binding. I added
a HttpBindingFactory that is able to return the correct binding helper
based on the method of the request that was received by the application.

Several tests have been provided to cover the newly created code.

Related to: https://www.pivotaltracker.com/story/show/149489911