Distinguish signature format and validation errors
When a signature can not be validated an exception is thrown. When a
signature could not be found or signature data was incomplete in the
message, the 'verify' methods return false. This was very misleading
in the error message and code so this commit changes the error message
and code comments to make this clear.