Don't track the composer.lock file
Adding a lock file to VCS makes more sense for applications than for
libraries. The lock file has no effect on the projects that require
this bundle. Not having a lock file has the benefit that we can build
on different versions of PHP with the latest framework versions
supported for that platform.

This also allows for building on future versions of PHP while keeping
support for older versions (like PHP<5.6 or future versions of PHP and
PHP5.6/7.x).

See: https://getcomposer.org/doc/02-libraries.md#lock-file