Add support for ignoring unknown SAML attributes
Prior to this change a SAML assertion could fail when the IdP returned
unknown attributes. This leads in an unhandled exception which resulted
in an application/php error.

By enabling the configuration directive "ignore_unknown_attributes" in the
hosted->attribute_dictionary unknown attibutes are ignored. The SAML assertion
will continue and only the values of the known attributes are available.
See README.md for the configuration details.
The configuration defaults to the previous behavior in which
an assertion fails when encountering an unknown SAML attribute.

(cherry picked from commit d0a322f)