Support multiple remote identity providers
This change allows configuring multiple remote identity providers
instead of a single remote identity provider.

This change is fully backwards compatible:

 - if a single identity provider is configured, the identity provider
   is still availble using the `surfnet_saml.remote.idp` service ID

 - if multiple identity providers are configured, a new repository is
   injected into the service container in identical fashion to how
   this works for service providers

The README is updated with instructions for both use-cases.

Ideally we'd ship this change with some tests, but this requires
writing an integration test for the bundle: the bulk of the change is
in how configuration is parsed. We don't yet have that kind of tests.