Add a SAML Authenticator
This authenticator is compatible with the Symfony Security system. It
creates a custom Authenticator class. The user must integrate this
authenticator in the application. And a custom user provider must be
created as this ties into the apolication logic to strongly to
facilitate this in the bundle.

Examples where added to EXAMPLES.md and README.md