Check destination in request URI Redirect binding
This check was kind of implemented but would always pass inspection as a
faulty logical expression was used.

The new situation: when a signed message is processed the Destination
MUST be present in hte SAML message. If not, it now exits with a new
error condition.

When the request uri does not start with the destination, the original
exception is thrown.

When dealing with unsigned requests. The destination is not mandatory.
But if it is set. We verify that the request uri starts with the destination.

Note this was only changed for Redirect binding. POST binding shows
similar faulty logic. That is tackeled in the next commit