Change getFullRequestUri strategy
The way SF parses and rewrites the REQUEST_URI maked me change the way
we constitute the fullRequestUri. SF will overwrite the REQUEST_URI when
the request uri is retrieved. That makes the request way safer, and more
predictable for the framework. However, we do need the full path
including the schema and hostname.

This should yield the most optimal result.